tech: Create services for buyer information

+++ db/migrate/20210115123707_create_order.rb
@@ -1,8 +1,8 @@
 class CreateOrder < ActiveRecord::Migration[6.1]
   def change
     create_table :orders do |t|
-      t.integer :external_code
-      t.integer :store_id
+      t.bigint :external_code
+      t.bigint :store_id
       t.timestamp :date_created
       t.timestamp :date_closed
       t.timestamp :last_updated
