+++ db/migrate/20210115184104_create_neighborhood.rb
@@ -2,7 +2,7 @@ class CreateNeighborhood < ActiveRecord::Migration[6.1]
   def change
     create_table :neighborhoods do |t|
       t.references :city, null: false, foreign_key: true
-      t.integer :external_code
+      t.bigint :external_code
       t.string :name
 
       t.timestamps
